% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Query OSV for package vulnerabilities}
\usage{
check_packages(check_data)
}
\arguments{
\item{check_data}{Data frame containing `Package` and `Version` columns.}
}
\value{
A data table of vulnerabilities and severity information, or `NULL`.
}
\description{
Query OSV for package vulnerabilities
}
\keyword{internal}
