% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_version.R
\name{compare_version}
\alias{compare_version}
\title{Compare a package version with a constraint}
\usage{
compare_version(x, y)
}
\arguments{
\item{x}{Package version to compare.}

\item{y}{Version constraint using `>`, `>=`, `<`, `<=`, `=` or `==`.}
}
\value{
A logical value indicating whether `x` satisfies `y`.
}
\description{
Compare a package version with a constraint
}
\keyword{internal}
