% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_body_list.R
\name{create_body_list}
\alias{create_body_list}
\title{Create a batch OSV query body}
\usage{
create_body_list(pkg = NULL, version = NULL)
}
\arguments{
\item{pkg}{Character vector of package names.}

\item{version}{Optional character vector of package versions.}
}
\value{
A list suitable for the OSV batch query endpoint, or `NULL`.
}
\description{
Create a batch OSV query body
}
\keyword{internal}
