% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cvss31.R
\name{cvss31_score}
\alias{cvss31_score}
\title{Calculate a CVSS 3.1 base score}
\usage{
cvss31_score(vector)
}
\arguments{
\item{vector}{A CVSS 3.1 vector string.}
}
\value{
A numeric score between 0 and 10.
}
\description{
Calculate a CVSS 3.1 base score
}
\keyword{internal}
