% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cveid.R
\name{get_cveid}
\alias{get_cveid}
\title{Retrieve upstream identifiers and fixed versions}
\usage{
get_cveid(resc_dt)
}
\arguments{
\item{resc_dt}{Data table of OSV result identifiers.}
}
\value{
A data table linking packages to upstream vulnerability identifiers.
}
\description{
Retrieve upstream identifiers and fixed versions
}
\keyword{internal}
