% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_severity.R
\name{get_severity}
\alias{get_severity}
\title{Retrieve CVSS severity for vulnerability identifiers}
\usage{
get_severity(ids)
}
\arguments{
\item{ids}{Character vector of upstream vulnerability identifiers.}
}
\value{
A data table with identifiers, CVSS scores, and severity labels.
}
\description{
Retrieve CVSS severity for vulnerability identifiers
}
\keyword{internal}
