% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cvss31.R
\name{parse_cvss31}
\alias{parse_cvss31}
\title{Parse a CVSS 3.1 vector}
\usage{
parse_cvss31(vector)
}
\arguments{
\item{vector}{A CVSS 3.1 vector string.}
}
\value{
A list containing the original vector, numeric score, and severity.
}
\description{
Parse a CVSS 3.1 vector
}
\keyword{internal}
