% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_library.R
\name{parse_library}
\alias{parse_library}
\title{Parse installed packages from library paths}
\usage{
parse_library(lib = .libPaths())
}
\arguments{
\item{lib}{Character vector of existing library paths.}
}
\value{
A data frame of installed packages and their source paths.
}
\description{
Parse installed packages from library paths
}
\keyword{internal}
