% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_renv_lock.R
\name{parse_renv_lock}
\alias{parse_renv_lock}
\title{Parse package records from renv lock files}
\usage{
parse_renv_lock(renv_file = "renv.lock")
}
\arguments{
\item{renv_file}{Character vector of paths to `renv.lock` files.}
}
\value{
A data frame containing package names, versions, source, and path.
}
\description{
Parse package records from renv lock files
}
\keyword{internal}
