% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_query.R
\name{set_query}
\alias{set_query}
\title{Build an OSV package query}
\usage{
set_query(pkg, version = NULL, eco = "CRAN")
}
\arguments{
\item{pkg}{Package name.}

\item{version}{Optional package version.}

\item{eco}{Package ecosystem, defaulting to `CRAN`.}
}
\value{
A list representing one OSV query.
}
\description{
Build an OSV package query
}
\keyword{internal}
