% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_packages.R
\name{normalize_packages}
\alias{normalize_packages}
\title{Normalize package data}
\usage{
normalize_packages(x)
}
\arguments{
\item{x}{Data frame with `Package` and `Version` columns.}
}
\value{
A normalized data frame with `Package`, `Version`, `source`, and `path` columns.
}
\description{
Normalize package data
}
\keyword{internal}
